Minor fixes for x64 compiler

+++ src/asm/libc/vfprintf.c
@@ -269,7 +269,7 @@ int vfprintf (FILE *fp, const char *str, va_list ap)
     // Write litteral characters
     if (*str != '%') {
       mxs = strchr(str, '%');
-      lg = (mxs == NULL) ? strlen(str) : (int)(mxs - str);
+      lg = (mxs == NULL) ? (int)strlen(str) : (int)(mxs - str);
       if (fp->write_(fp, str, lg) < 0)
         return -1;
